backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
torch
pydantic-settings
python-dotenv